Get unsafe.ArbitraryType (kinda) working, wrap more output-by-fn types
While this builds and passes tests, the unsafe.* stuff is still a
kludge: unsafe.ArbitraryType is now processed, but special-cased in
various places. Similarly, unsafe.Offsetof is simply not generated, as
it's really more of a syntactic operator in Go, and supporting a Joker
equivalent (if even possible) is not within the scope of this current
work.