Avoid other issues while working around lack of type aliases
Joker (maybe Clojure generally) presents challenges doing things like `(catch Error ...)`, where `Error` is a builtin type, after having done a `(use ...)` or `:use ...` that defines the `Error` symbol for the current namespace.

So, rather than rely on catching an error trying to bring in the new `go.std.io.fs` library, look for it in `(all-ns)` and, if not present, substitute `go.std.os`, and bring that in (which is redundant due to bringing it in for `LinkError`).