Generate calls to new FieldAs_ns_* routines for building structs
This might someday work, but for now it runs into the problem that this new code wants to be sure the FieldAs_ns_* routine in question actually exists, but this code is usually run when generating ctors, which happens before those routines are generated.

In particular, genTypeFromDb happens for all types before any GenType calls are made. The former generate the ctors, while the latter generate the support routines they need.

So, this needs to be solved...as expected awhile back.