[WiP] Try a new approach to supporting more methods across ptr/val issues
This breaks DotNode and NilNode (in text/template/parse) again, but now it's clearer to me what's different about them.

* DotNode embeds NodeType
* NodeType implements Type()
* But *DotNode overrides Type(), so that’s what gets invoked (semantically), even when the operand is DotNode instead of *DotNode
* Since o.O.(parse.DotNode) is not addressable, the Go compiler rejects it

So, while all the other methods that are enabled by this commit seem to be okay, the Type() methods enabled for *DotNode and *NilNode, though they look innocuous because `NodeType Type()` results from the embedding, in fact result in `*Type()` methods being denoted at compile time, which fails for `DotNode` (vs `*DotNode`) method generation.

The solution seems to be to not generate an embedded method if there's an incompatible override of it.